+++ src/EmbeddedManifold.jl
@@ -43,7 +43,7 @@ get_embedding_type(::EmbeddedManifold) = EmbeddedManifoldType()
 
 function allocate_result(M::EmbeddedManifold, f::typeof(project), x...)
     T = allocate_result_type(M, f, x)
-    return allocate(M, x[1], T, representation_size(base_manifold(M)))
+    return allocate(M, x[1], T, representation_size(base_manifold(M, T)))
 end
 
 """