+++ src/TangentSpace.jl
@@ -239,9 +239,9 @@ function Random.rand!(rng::AbstractRNG, TpM::TangentSpace, X; vector_at = nothin
     return X
 end
 
-function representation_size(TpM::TangentSpace)
-    return representation_size(TpM.manifold)
-end
+representation_size(TpM::TangentSpace) = representation_size(TpM.manifold)
+representation_size(TpM::TangentSpace, T::Type) = representation_size(TpM.manifold, T)
+
 
 function Base.show(io::IO, ::MIME"text/plain", TpM::TangentSpace)
     println(io, "Tangent space to the manifold $(base_manifold(TpM)) at point:")